+++ flower/src/main/java/com/flower/repository/FavoriteRepository.java
@@ -6,20 +6,20 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import java.util.List;
 import java.util.Optional;
 
-public interface FavoriteRepository extends JpaRepository<Favorites, Long> {
+public interface FavoriteRepository extends JpaRepository<Favorites, Integer> {
 
     // 사용자의 즐겨찾기 목록 조회
-    List<Favorites> findByUserIdOrderByCreatedAtDesc(Long userId);
+    List<Favorites> findByUserIdOrderByCreatedAtDesc(Integer userId);
 
     // 특정 꽃이 즐겨찾기에 있는지 확인
-    Optional<Favorites> findByUserIdAndFlowerId(Long userId, Long flowerId);
+    Optional<Favorites> findByUserIdAndFlowerId(Integer userId, Integer flowerId);
 
     // 즐겨찾기 존재 여부 확인
-    boolean existsByUserIdAndFlowerId(Long userId, Long flowerId);
+    boolean existsByUserIdAndFlowerId(Integer userId, Integer flowerId);
 
     // 특정 꽃 즐겨찾기 삭제
-    void deleteByUserIdAndFlowerId(Long userId, Long flowerId);
+    void deleteByUserIdAndFlowerId(Integer userId, Integer flowerId);
 
     // 사용자의 즐겨찾기 개수
-    long countByUserId(Long userId);
+    long countByUserId(Integer userId);
 }
